cpu: change adc_sample return to int32_t

+++ cpu/cc2538/periph/adc.c
@@ -51,7 +51,7 @@ int adc_init(adc_t line)
     return 0;
 }
 
-int adc_sample(adc_t line, adc_res_t res)
+int32_t adc_sample(adc_t line, adc_res_t res)
 {
     /* check if adc line valid */
     if (line >= ADC_NUMOF) {
@@ -112,5 +112,5 @@ int adc_sample(adc_t line, adc_res_t res)
         sample = 0;
     }
 
-    return (int)sample;
+    return sample;
 }
